requirements.txt
attrs
dataclasses
python-box
thingy

line_profiler
